Release/05 10 2019

+++ app/src/main/java/com/future/medan/backend/models/entity/Product.java
@@ -1,14 +1,13 @@
 package com.future.medan.backend.models.entity;
 
-import com.future.medan.backend.models.constants.ProductConstant;
+import com.future.medan.backend.constants.ProductConstant;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-//import org.hibernate.annotations.Fetch;
 
 import javax.persistence.*;
-import java.util.Set;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -28,7 +27,7 @@ public class Product extends BaseEntity {
     private String description;
 
     @Column(name = ProductConstant.PRODUCT_PRICE)
-    private Float price;
+    private BigDecimal price;
 
     @Column(name = ProductConstant.PRODUCT_IMAGE)
     private String image;
